infrastructure/expected-fail/assert_implements.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>assert_implements failure values</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_implements(false, 'assert implements 1');
}, 'boolean false value');

test(() => {
  assert_implements(0, 'assert implements 1');
}, 'falsy number value');

test(() => {
  assert_implements('', 'assert implements 1');
}, 'falsy string value');

test(() => {
  assert_implements(null, 'assert implements 1');
}, 'falsy null value');

test(() => {
  assert_implements(undefined, 'assert implements 1');
}, 'falsy undefined value');

test(() => {
  assert_implements(NaN, 'assert implements 1');
}, 'falsy NaN value');
</script>
